Seperate local storage string and object examples
These have been seperated into two different components

- Fix example code in README not importing from package.
