+++ examples/src/App.tsx
@@ -1,11 +1,13 @@
 import "./App.css";
-import ComponentOne from "./components/ComponentOne";
+import ComponentOne from "./components/component_one/ComponentOne";
+import ComponentTwo from "./components/component_two/ComponentTwo";
 
 function App() {
   return (
     <div>
       <h1>useLocalStorage test updates</h1>
       <ComponentOne />
+      <ComponentTwo />
     </div>
   );
 }
