+++ examples/src/components/component_one/ComponentOne.tsx
@@ -1,0 +1,38 @@
+import useLocalStorage from "react-persist-local-storage";
+import "./ComponentOne.css";
+
+/**
+ * This example component supports a local storage value of object, string
+ * and null types.
+ */
+const ComponentOne = () => {
+  const key = "component.one.example";
+  const [value, setValue, deleteValue] = useLocalStorage(
+    key,
+    "Test Value"
+  );
+
+  return (
+    <div className="component-border" style={{ paddingBottom: "15px" }}>
+      <h3>Component One</h3>
+      <input
+        value={value || ""}
+        onChange={(e) => {
+          setValue(e.target.value);
+        }}
+      />
+      <div>
+        <p>
+          <b>Current local storage value:</b>
+        </p>
+        <p>{value}</p>
+      </div>
+
+      <button disabled={!value} onClick={() => deleteValue(key)}>
+        Delete "{key}"
+      </button>
+    </div>
+  );
+};
+
+export default ComponentOne;
