+++ examples/src/components/component_two/ComponentTwo.css
@@ -1,0 +1,12 @@
+.component-border {
+  border: 1px solid rgb(105, 105, 105);
+}
+
+button[disabled="disabled"],
+button:disabled {
+  border: 1px solid #999999;
+  background-color: #575757;
+  color: #878787;
+  cursor: not-allowed;
+  pointer-events: none;
+}
